bbdd/requirements.txt
fastapi==0.115.6
uvicorn==0.23.1
torch @ https://download.pytorch.org/whl/cu121/torch-2.5.1%2Bcu121-cp311-cp311-linux_x86_64.whl
pandas==2.2.2
chronos-forecasting==1.4.1
python-multipart